chore: refactor user seeder
