+++ test/seed/index.ts
@@ -18,3 +18,4 @@
 
 export { default as RbacSeeder } from './rbac';
 export { default as SellerPayoutSeeder } from './seller-payout';
+export { default as UserSeeder } from './user';
