+++ test/unit/controller/event-shift-controller.ts
@@ -24,7 +24,7 @@ import EventShift from '../../../src/entity/event/event-shift';
 import EventShiftAnswer from '../../../src/entity/event/event-shift-answer';
 import AssignedRole from '../../../src/entity/rbac/assigned-role';
 import Database from '../../../src/database/database';
-import { seedEvents, seedUsers } from '../../seed-legacy';
+import { seedEvents } from '../../seed-legacy';
 import TokenHandler from '../../../src/authentication/token-handler';
 import Swagger from '../../../src/start/swagger';
 import { json } from 'body-parser';
@@ -44,7 +44,7 @@ import Event, { EventType } from '../../../src/entity/event/event';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
 import Role from '../../../src/entity/rbac/role';
-import { RbacSeeder } from '../../seed';
+import { RbacSeeder, UserSeeder } from '../../seed';
 
 describe('EventShiftController', () => {
   let ctx: {
@@ -87,7 +87,7 @@ describe('EventShiftController', () => {
     await User.save(adminUser);
     await User.save(localUser);
 
-    const users = await seedUsers();
+    const users = await new UserSeeder().seedUsers();
     const { roleAssignments, events, eventShifts, eventShiftAnswers } = await seedEvents(users);
 
     // start app
