+++ test/unit/gewis/gewisdb.ts
@@ -20,13 +20,13 @@ import sinon, { SinonSandbox, SinonSpy } from 'sinon';
 import { defaultBefore, DefaultContext, finishTestDB } from '../../helpers/test-helpers';
 import User from '../../../src/entity/user/user';
 import GewisUser from '../../../src/gewis/entity/gewis-user';
-import { seedUsers } from '../../seed-legacy';
 import seedGEWISUsers from '../../../src/gewis/database/seed';
 import GewisDBService from '../../../src/gewis/service/gewisdb-service';
 import { BasicApi, MemberAllAttributes, MembersApi } from 'gewisdb-ts-client';
 import nodemailer, { Transporter } from 'nodemailer';
 import Mailer from '../../../src/mailer';
 import { In } from 'typeorm';
+import { UserSeeder } from '../../seed';
 
 describe('GEWISDB Service', () => {
 
@@ -45,7 +45,7 @@ describe('GEWISDB Service', () => {
     ctx = {
       ...(await defaultBefore()),
     } as any;
-    ctx.users = await seedUsers();
+    ctx.users = await new UserSeeder().seedUsers();
     ctx.gewisUsers = await seedGEWISUsers(ctx.users);
 
     Mailer.reset();
