+++ test/unit/service/event-service.ts
@@ -19,7 +19,7 @@
 import { describe } from 'mocha';
 import { Connection } from 'typeorm';
 import User from '../../../src/entity/user/user';
-import { seedEvents, seedUsers } from '../../seed-legacy';
+import { seedEvents } from '../../seed-legacy';
 import Event, { EventType } from '../../../src/entity/event/event';
 import EventShift from '../../../src/entity/event/event-shift';
 import EventShiftAnswer, { Availability } from '../../../src/entity/event/event-shift-answer';
@@ -34,6 +34,7 @@ import nodemailer, { Transporter } from 'nodemailer';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
 import Role from '../../../src/entity/rbac/role';
+import { UserSeeder } from '../../seed';
 
 describe('eventService', () => {
   let ctx: {
@@ -50,7 +51,7 @@ describe('eventService', () => {
     const connection = await Database.initialize();
     await truncateAllTables(connection);
 
-    const users = await seedUsers();
+    const users = await new UserSeeder().seedUsers();
     const { roleAssignments, events, eventShifts: allEventShifts, eventShiftAnswers } = await seedEvents(users);
 
     const eventShifts = allEventShifts.filter((s) => s.deletedAt == null);
