+++ test/unit/service/rbac-service.ts
@@ -23,11 +23,10 @@ import { DataSource } from 'typeorm';
 import { SeededRole } from '../../seed/rbac';
 import User, { UserType } from '../../../src/entity/user/user';
 import database from '../../../src/database/database';
-import { seedUsers } from '../../seed-legacy';
 import { finishTestDB } from '../../helpers/test-helpers';
 import { UpdateRoleRequest } from '../../../src/controller/request/rbac-request';
 import Role from '../../../src/entity/rbac/role';
-import { RbacSeeder } from '../../seed';
+import { RbacSeeder, UserSeeder } from '../../seed';
 
 const all = { all: new Set<string>(['*']) };
 const own = { own: new Set<string>(['*']) };
@@ -42,7 +41,7 @@ describe('RBACService', () => {
 
   before(async () => {
     const connection = await database.initialize();
-    const users = await seedUsers();
+    const users = await new UserSeeder().seedUsers();
 
     const roles = await new RbacSeeder().seedRoles([{
       name: 'system-default-role',
