+++ test/unit/service/stripe-service.ts
@@ -21,7 +21,7 @@ import { expect } from 'chai';
 import Stripe from 'stripe';
 import User from '../../../src/entity/user/user';
 import Database, { AppDataSource } from '../../../src/database/database';
-import { seedStripeDeposits, seedUsers } from '../../seed-legacy';
+import { seedStripeDeposits } from '../../seed-legacy';
 import StripeDeposit from '../../../src/entity/stripe/stripe-deposit';
 import StripeService, { STRIPE_API_VERSION } from '../../../src/service/stripe-service';
 import DineroTransformer from '../../../src/entity/transformer/dinero-transformer';
@@ -30,6 +30,7 @@ import BalanceResponse from '../../../src/controller/response/balance-response';
 import { StripeRequest } from '../../../src/controller/request/stripe-request';
 import { truncateAllTables } from '../../setup';
 import { finishTestDB } from '../../helpers/test-helpers';
+import { UserSeeder } from '../../seed';
 
 describe('StripeService', async (): Promise<void> => {
   let shouldSkip: boolean;
@@ -51,7 +52,7 @@ describe('StripeService', async (): Promise<void> => {
     const connection = await Database.initialize();
     await truncateAllTables(connection);
 
-    const users = await seedUsers();
+    const users = await new UserSeeder().seedUsers();
     const { stripeDeposits } = await seedStripeDeposits(users);
 
     const stripeService = new StripeService();
