+++ test/unit/service/transaction-service.ts
@@ -32,7 +32,6 @@ import {
   seedProductCategories,
   seedProducts,
   seedTransactions,
-  seedUsers,
   seedVatGroups,
 } from '../../seed-legacy';
 import TransactionService, { TransactionFilterParameters } from '../../../src/service/transaction-service';
@@ -55,6 +54,7 @@ import { createInvoiceWithTransfers } from './invoice-service';
 import { truncateAllTables } from '../../setup';
 import ProductRevision from '../../../src/entity/product/product-revision';
 import { calculateBalance } from '../../helpers/balance';
+import { UserSeeder } from '../../seed';
 
 chai.use(deepEqualInAnyOrder);
 
@@ -80,7 +80,7 @@ describe('TransactionService', (): void => {
     await truncateAllTables(connection);
 
     const app = express();
-    const users = await seedUsers();
+    const users = await new UserSeeder().seedUsers();
     const vatGropus = await seedVatGroups();
     const categories = await seedProductCategories();
     const { productRevisions } = await seedProducts(users, categories, vatGropus);
